Fixes from Francois and Marc PR reviews
