+++ src/pages/EnablePayments/OnfidoStep.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import {withOnyx} from 'react-native-onyx';
+import lodashGet from 'lodash/get';
 import Onfido from '../../components/Onfido';
 import ONYXKEYS from '../../ONYXKEYS';
 import * as BankAccounts from '../../libs/actions/BankAccounts';
@@ -32,6 +33,17 @@ const defaultProps = {
 };
 
 class OnfidoStep extends React.Component {
+    componentDidMount() {
+        // Once in Onfido step, if we somehow don't have the personal info, go back to previous step, as we need them for Onfido{
+        const firstName = lodashGet(this.props, 'walletAdditionalDetailsDraft.legalFirstName');
+        const lastName = lodashGet(this.props, 'walletAdditionalDetailsDraft.legalLastName');
+        const dob = lodashGet(this.props, 'walletAdditionalDetailsDraft.dob');
+
+        if (!firstName || !lastName || !dob) {
+            Wallet.updateCurrentStep(CONST.WALLET.STEP.ADDITIONAL_DETAILS);
+        }
+    }
+
     /**
      * @returns {boolean|*}
      */
